integration-and-sync-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
pydantic==2.5.3
pydantic-settings==2.0.3
atlassian-python-api==3.41.0
